requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic[email]==2.7.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
cryptography==42.0.5
python-multipart==0.0.9
pyjwt==2.8.0
alembic==1.13.1
jinja2==3.1.3
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
stripe==8.7.0
requests==2.31.0
web3==6.15.0
apscheduler==3.11.0
aiofiles==23.2.1
